linuxstacks refactored
